+++ R/sim_mixed_cc.R
@@ -8,6 +8,7 @@
 #' @param sub_sd the SD of subject random intercepts
 #' @param item_sd the SD of item random intercepts
 #' @param error_sd the SD of the error term
+#' @param seed a single value, interpreted as an integer, or NULL (see set.seed)
 #'
 #' @return a tbl 
 #' @export
@@ -15,18 +16,40 @@
 #' @examples
 #' 
 #' sim_mixed_cc(10, 10)
-sim_mixed_cc <- function(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1, item_sd = 1, error_sd = 1) {
+sim_mixed_cc <- function(sub_n = 100, item_n = 20, grand_i = 0, 
+                         sub_sd = 1, item_sd = 1, error_sd = 1, seed = NULL) {
+  set.seed(seed)
+  
   # sample subject random intercepts -------------------------------------------
-  new_sub <- data.frame(
-    sub_id = make_id(1:sub_n, "S"),
-    sub_i = stats::rnorm(sub_n, 0, sub_sd)
-  )
+  if (length(sub_sd) == sub_n) {
+    # use exact subject intercepts and names
+    new_sub <- data.frame(
+      sub_id = rownames(sub_sd),
+      sub_i = sub_sd %>% unname()
+    )
+  } else if (length(sub_sd) == 1) {
+    # sample new subjects and intercepts
+    new_sub <- data.frame(
+      sub_id = make_id(1:sub_n, "S"),
+      sub_i = stats::rnorm(sub_n, 0, sub_sd)
+    )
+  }
+  
   
   # sample item random intercepts ----------------------------------------------
-  new_item <- data.frame(
-    item_id = make_id(1:item_n, "I"),
-    item_i = stats::rnorm(item_n, 0, item_sd)
-  )
+  if (length(item_sd) == item_n) {
+    # use exact item intercepts and names
+    new_item <- data.frame(
+      item_id = rownames(item_sd),
+      item_i = item_sd %>% unname()
+    )
+  } else if (length(item_sd) == 1) {
+    # sample new items and intercepts
+    new_item <- data.frame(
+      item_id = make_id(1:item_n, "I"),
+      item_i = stats::rnorm(item_n, 0, item_sd)
+    )
+  }
   
   new_obs <- expand.grid(
     sub_id = new_sub$sub_id,
@@ -38,7 +61,8 @@ sim_mixed_cc <- function(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1, item
       grand_i = grand_i,
       err = stats::rnorm(nrow(.), 0, error_sd),
       val = grand_i + sub_i + item_i + err
-    )
+    ) %>%
+    dplyr::select(sub_id, item_id, val, grand_i, sub_i, item_i, err)
   
   new_obs
 }